requirements.txt
numpy
matplotlib
torch 
pillow
scikit-learn
